requirements.txt
streamlit
torch
torchvision
pillow
numpy
pandas
matplotlib
seaborn
scikit-learn
gdown
